+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a CodeEditor component based on CodeMirror 5
 - Add the ability to enable/disable plugins by URL
 - Add support for plugins stored (not just cached) locally in indexeddb
+- Add the ability to create/edit plugins by writing their source code inside Asset Link
 
 ### Changed